move quality constants into pre-existent class
